moved orderItem to components and rednder styles depending on page
