+++ client/src/components/OrderItem/FavoriteSx/crossSx.js
@@ -1,0 +1,6 @@
+export const favCrossSx = {
+	display: "flex",
+	flexDirection: "column",
+	justifyContent: "center",
+	alignItems: "center",
+};
