+++ client/src/components/OrderItem/FavoriteSx/priceSx.js
@@ -1,0 +1,1 @@
+export const favPriceSX = { display: "flex", flexDirection: "column", justifyContent: "center" };
