+++ client/src/components/OrderItem/OrderItem.scss
@@ -1,4 +1,5 @@
-@import "../../../boilerplates/styles/variables";
+// @import "../../../boilerplates/styles/variables";
+@import "../../boilerplates/styles//variables";
 
 .item-product {
 	display: flex;
